BufferPolylineCollection: Relocate clamped renderer into collection
